agents/architect/requirements.txt
functions-framework==3.*
Flask==3.0.3
gunicorn==21.2.0
requests==2.32.3
google-api-python-client==2.145.0
google-auth==2.29.0
google-cloud-vertexai==1.55.0
sendgrid==6.11.0
PyGithub==2.3.0
